pyproject.toml
[project]
name = "sql-dojo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-limiter>=3.12",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "supabase>=2.0.0",
]
